fix: resolve warnings and missing TreeNode struct definition

+++ String/caseConv.cpp
@@ -3,13 +3,13 @@ int main(){
     std::string str = "xecgvux fhijoz";
 
     //convert to uppercase
-    for(int i=0;i<str.size();i++){
+    for(size_t i=0;i<str.size();i++){
 str[i]-=32;
     }
     std::cout<<str<<std::endl;
 
     //convert to lowercase
-    for(int i=0;i<str.size();i++){
+    for(size_t i=0;i<str.size();i++){
 str[i]+=32;
     }
     std::cout<<str<<std::endl;
